Fixing warnings on 32bit platform for Fmd/Stringconversion functions. Fixing printout warning in eosfstcp. Disabling assembler code for 32bit platform in crc32c.

+++ fst/txqueue/eosfstcp.cc
@@ -847,7 +847,7 @@ int main(int argc, char* argv[]) {
       }
       
       if (nwrite != nread) {
-	fprintf(stderr, "error: write failed on destination file %s - wrote %ld/%ld bytes - destination file is incomplete!\n",destination[i],nwrite,nread);
+	fprintf(stderr, "error: write failed on destination file %s - wrote %lld/%lld bytes - destination file is incomplete!\n",destination[i],(long long)nwrite,(long long)nread);
 	exit(-EIO);
       }
